refactor: move canaryinterface to canary directory
