+++ src/components/Canary/CanaryInterface.tsx
@@ -2,15 +2,11 @@ import { orderBy } from "lodash";
 import React, { useEffect, useMemo, useState } from "react";
 import { useSearchParams } from "react-router-dom";
 import { HealthCheck } from "../../api/types/health";
-import { CanaryTabs, filterChecksByTabSelection } from "../Canary/CanaryTabs";
-import { ChecksListing } from "../Canary/ChecksListing";
-import { CanarySorter } from "../Canary/data";
-import { filterChecks, filterChecksByText } from "../Canary/filter";
-import {
-  filterChecksByLabels,
-  getLabelFilters,
-  getLabels
-} from "../Canary/labels";
+import { CanaryTabs, filterChecksByTabSelection } from "./CanaryTabs";
+import { ChecksListing } from "./ChecksListing";
+import { CanarySorter } from "./data";
+import { filterChecks, filterChecksByText } from "./filter";
+import { filterChecksByLabels, getLabelFilters, getLabels } from "./labels";
 
 type Props = {
   checks?: HealthCheck[];
